+++ analysis/variability_analysis.py
